firstExpression 파라미터 전달 방식을 Nullable 타입 체크 방식으로 변경

+++ MSProgrammerCalculator/MSProgrammerCalculator/ViewModels/CalculatorViewModel.cs
@@ -111,7 +111,7 @@ namespace MSProgrammerCalculator.ViewModels
                 _currentOperandChanged = false;
 
                 NumericalExpression = _currentContext.Expression;
-                DisplayValue = _currentContext.Result;
+                DisplayValue = (long)_currentContext.Result;
             }
         }
 
@@ -163,7 +163,7 @@ namespace MSProgrammerCalculator.ViewModels
             _calculator.Evaluate();
 
             NumericalExpression = _currentContext.Expression;
-            DisplayValue = _currentContext.Result;
+            DisplayValue = (long)_currentContext.Result;
         }
 
         private void BaseNumberChanged()
